+++ src/actinia_core/resources/location_management.py
@@ -33,7 +33,7 @@ import shutil
 import pickle
 from flask_restful_swagger_2 import swagger, Schema
 from .common.app import auth
-from .common.logging_interface import log_api_call
+from .common.api_logger import log_api_call
 from .user_auth import very_admin_role
 from .user_auth import check_user_permissions
 from .common.response_models import ProcessingResponseModel
