+++ src/actinia_core/resources/user_api_key.py
@@ -35,7 +35,7 @@ from flask_restful_swagger_2 import Schema, swagger
 from flask import jsonify, make_response, g
 from .base_login import LoginBase
 from .common.app import auth
-from .common.logging_interface import log_api_call
+from .common.api_logger import log_api_call
 
 
 __license__ = "GPLv3"
